archive/python/requirements.dev.txt
# Dev/test tooling (installed only in the Docker dev target)
pytest>=7.0.0,<9.0.0
pytest-cov>=4.1.0,<5.0.0
responses>=0.25.0,<1.0.0
hypothesis>=6.98.0,<7.0.0

# Formatting / linting / typing
ruff
black
mypy
pandas-stubs
types-PyYAML
types-Markdown

# Build tooling for the Rust extension
maturin>=1.0,<2.0
